Coding/flight-monitor/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
google-auth>=2.23.0
google-api-python-client>=2.100.0
apscheduler>=3.10.0
